requirements.txt
python-telegram-bot
fastapi==0.111.0
uvicorn==0.23.2